requirements.txt
tensorflow==2.11.0
keras==2.11.0
numpy==1.23.5
pillow==9.5.0
rembg==2.0.50
onnxruntime==1.15.1
python-telegram-bot==20.3
numba==0.57.1
tbb==2021.10.0
